refactor: 모달 관련 타입만 분리
