+++ hooks/auth/useGoogleLoginMutation.ts
@@ -15,7 +15,7 @@ import { COOKIEOPTIONS_ACCESS, COOKIEOPTIONS_REFRESH } from '@/constant/cookieOp
 const redirectUri =
   typeof window !== 'undefined'
     ? `${window.location.origin}/auth/callback`
-    : process.env.NEXT_PUBLIC_REDIRECT_URI || '';
+    : process.env.NEXT_PUBLIC_REDIRECT_URI || 'http://localhost:3000/auth/callback';
 
 async function googleLogin(authorizationCode: string): Promise<OAuthLoginResponse> {
   const response = await authApi.get('/auth/google', {
