+++ stores/modalStore.ts
@@ -1,55 +1,9 @@
 'use client';
 
-import { NoteConfirmModalProps } from '@/types/note.type';
 import { createStore } from 'zustand/vanilla';
 
-export type EditModalProps = {
-  onConfirm: (value: string) => void;
-  onCancel: () => void;
-  initialValue: string;
-};
-
-export type DeleteModalProps = {
-  onConfirm: () => void;
-  onCancel: () => void;
-};
-
-export type NoteLinkModalProps = {
-  defaultValue: string | undefined;
-  onSubmit: () => void;
-  linkInputRef: React.RefObject<HTMLInputElement | null>;
-};
-
-export interface ModalStore {
-  isIOSPWAGuideModalOpen: boolean;
-  isNoteConfirmModalOpen: boolean;
-  noteConfirmModalProps: NoteConfirmModalProps | null;
-  isTodoCreateModalOpen: boolean;
-  isTodoCreateCheckModalOpen: boolean;
-  isTodoDeleteConfirmModalOpen: boolean;
-  todoDeleteConfirmModalProps: DeleteModalProps | null;
-  isNoteLinkModalOpen: boolean;
-  noteLinkModalProps: NoteLinkModalProps | null;
-  isGoalDeleteModalOpen: boolean;
-  goalDeleteModalProps: DeleteModalProps | null;
-  isGoalEditModalOpen: boolean;
-  goalEditModalProps: EditModalProps | null;
-  isGoalCreateModalOpen: boolean;
-  isLoading: boolean;
-
-  setIOSPWAGuideModalOpen: (isOpen: boolean) => void;
-  setIsNoteConfirmModalOpen: (isOpen: boolean, props?: NoteConfirmModalProps) => void;
-  setIsTodoCreateModalOpen: (isOpen: boolean) => void;
-  setIsTodoCreateCheckModalOpen: (isOpen: boolean) => void;
-  setIsTodoDeleteConfirmModalOpen: (isOpen: boolean, props?: DeleteModalProps) => void;
-  setNoteLinkModalOpen: (isOpen: boolean, props?: NoteLinkModalProps) => void;
-  setGoalDeleteModalOpen: (isOpen: boolean, props?: DeleteModalProps) => void;
-  setGoalEditModalOpen: (isOpen: boolean, props?: EditModalProps) => void;
-  setIsGoalCreateModalOpen: (isOpen: boolean) => void;
-  setIsLoading: (isLoading: boolean) => void;
-}
-
-export const defaultInitState: ModalStore = {
+import { ModalStoreType } from '@/types/modalStore.types';
+export const defaultInitState: ModalStoreType = {
   isIOSPWAGuideModalOpen: false,
   isNoteConfirmModalOpen: false,
   noteConfirmModalProps: null,
@@ -78,8 +32,8 @@ export const defaultInitState: ModalStore = {
   setIsLoading: () => {}
 };
 
-export const createModalStore = (initState: Partial<ModalStore> = defaultInitState) => {
-  return createStore<ModalStore>()((set) => ({
+export const createModalStore = (initState: Partial<ModalStoreType> = defaultInitState) => {
+  return createStore<ModalStoreType>()((set) => ({
     ...defaultInitState,
     ...initState,
     setIOSPWAGuideModalOpen: (isOpen) => set({ isIOSPWAGuideModalOpen: isOpen }),
