+++ types/note.types.ts
@@ -1,4 +1,4 @@
-import type { TodoType } from './todo.type';
+import type { TodoType } from './todo.types';
 import type { ISODateString } from './ISODateString.type';
 import type { GoalSummary } from './goal.types';
 type NoteCommonFields = {
@@ -48,17 +48,4 @@ export type TempNoteType = {
 
 export type TempNoteContentType = {
   [goalId: number]: TempNoteType[];
-};
-
-export type NoteConfirmModalProps = {
-  type: 'temp' | 'delete';
-  contentTitle: string;
-  onCancel: () => void;
-  onConfirm: () => void;
-};
-
-export type NoteLinkModalProps = {
-  defaultValue: string | undefined;
-  onSubmit: () => void;
-  linkInputRef: React.RefObject<HTMLInputElement | null>;
 };
