+++ utils/dashboard/getBgColorForGoal.ts
@@ -1,4 +1,4 @@
-import type { TodoType } from '@/types/todo.type';
+import type { TodoType } from '@/types/todo.types';
 import type { GoalType } from '@/types/goal.types';
 
 export const getBgColorForGoal = (
