+++ utils/dashboard/getCompletedGoalCounts.ts
@@ -1,6 +1,6 @@
 import { getGoalTodos } from './goalUtil';
 
-import type { TodoType } from '@/types/todo.type';
+import type { TodoType } from '@/types/todo.types';
 
 export const getCompletedGoalCounts = (todos: TodoType[], year: number, month: number): number => {
   // 해당 월에 완료된 todos 필터링
