+++ utils/dashboard/getWeeklyCompletionData.ts
@@ -1,7 +1,7 @@
 import { calcWeeklyCompletionPercentages } from '../percentageUtils';
 import { getWeeklyCompletionCounts } from './getWeeklyCompletionCounts';
 
-import type { TodoType } from '@/types/todo.type';
+import type { TodoType } from '@/types/todo.types';
 
 export const getWeeklyCompletionData = (todos: TodoType[]) => {
   const doneTodos = todos.filter((todo) => todo.done === true);
