+++ utils/getKakaoLoginUrl.ts
@@ -1,6 +1,6 @@
 const getKakaoLoginUrl = () => {
   const clientId = process.env.NEXT_PUBLIC_KAKAO_CLIENT_ID;
-  const redirectUri = process.env.NEXT_PUBLIC_REDIRECT_URI;
+  const redirectUri = process.env.NEXT_PUBLIC_REDIRECT_URI || 'http://localhost:3000/auth/callback';
 
   if (!clientId || !redirectUri) {
     throw new Error('Kakao OAuth 환경 변수가 설정되지 않았습니다.');
